fixed issue 178 & 179
