+++ contracts/raffle-instance/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "raffle-instance"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+soroban-sdk = { workspace = true }
+raffle-shared = { path = "../raffle-shared" }
