+++ contracts/raffle-shared/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "raffle-shared"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["lib"]
+
+[dependencies]
+soroban-sdk = { workspace = true }
