+++ contracts/raffle/Cargo.toml
@@ -1,22 +1,24 @@
 [package]
-name = "hello-world"
-version = "0.0.0"
+name = "raffle-factory"
+version = "0.1.0"
 edition = "2021"
 publish = false
 
 [lib]
-crate-type = ["lib", "cdylib"]
+crate-type = ["cdylib"]
 doctest = false
 
 [dependencies]
 soroban-sdk = { workspace = true }
+raffle-shared = { path = "../raffle-shared" }
 
 [dev-dependencies]
 soroban-sdk = { workspace = true, features = ["testutils"] }
 proptest = "1"
 ed25519-dalek = "2"
+raffle-instance = { path = "../raffle-instance" }
 
-# Optimize WASM size - use same release profile as workspace
+# Optimize WASM size
 [profile.release]
 opt-level = "z"
 overflow-checks = true
@@ -25,4 +27,4 @@ strip = "symbols"
 debug-assertions = false
 panic = "abort"
 codegen-units = 1
-lto = "fat"  # Use "fat" LTO for maximum WASM size reduction
+lto = "fat"
