--- contracts/raffle/src/instance/test.rs
@@ -1,2930 +1,0 @@
-#![cfg(test)]
-#![allow(warnings)]
-#![allow(clippy::all)]
-
-use super::*;
-use crate::events::{RaffleFinalized, RandomnessType};
-use crate::types::PaginationParams;
-use crate::{ContractError, RaffleFactory, RaffleFactoryClient};
-use soroban_sdk::{
-    testutils::{Address as _, Events, Ledger},
-    token, Address, BytesN, Env, String, Symbol, TryFromVal,
-};
-#[contract]
-pub struct MockFactory;
-#[contractimpl]
-impl MockFactory {
-    pub fn record_volume(env: Env, token: Address, amount: i128) {}
-}
-
-#[contract]
-pub struct NativeMock;
-#[contractimpl]
-impl NativeMock {
-    pub fn name(env: Env) -> String {
-        String::from_str(&env, "native")
-    }
-    pub fn symbol(env: Env) -> String {
-        String::from_str(&env, "XLM")
-    }
-    pub fn decimals(env: Env) -> u32 {
-        7
-    }
-    pub fn transfer(env: Env, from: Address, to: Address, amount: i128) {}
-    pub fn balance(env: Env, owner: Address) -> i128 {
-        0
-    }
-    pub fn mint(env: Env, to: Address, amount: i128) {}
-}
-
-/// HELPER: Standardized environment setup
-fn setup_raffle_env(
-    env: &Env,
-    source: RandomnessSource,
-    oracle: Option<Address>,
-    fee_bp: u32,
-    treasury: Option<Address>,
-) -> (
-    ContractClient<'_>,
-    Address,
-    Address,
-    token::StellarAssetClient<'_>,
-    Address,
-    Address,
-) {
-    let creator = Address::generate(env);
-    let buyer = Address::generate(env);
-    let admin = Address::generate(env);
-    let factory_admin = Address::generate(env);
-
-    // Register factory as a dummy contract so env.as_contract works
-
-    let factory = env.register(MockFactory, ());
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-    let admin_client = token::StellarAssetClient::new(env, &token_id);
-
-    admin_client.mint(&creator, &1_000_000i128);
-    admin_client.mint(&buyer, &1_000_000i128);
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(env, &contract_id);
-
-    let mut prizes = Vec::new(env);
-    prizes.push_back(10000);
-
-    let config = RaffleConfig {
-        description: String::from_str(env, "Audit Raffle"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id,
-        prize_amount: 100_000i128,
-        prizes,
-        randomness_source: source,
-        oracle_address: oracle,
-        protocol_fee_bp: fee_bp,
-        treasury_address: treasury,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(env, &[1u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-
-    (client, creator, buyer, admin_client, factory, factory_admin)
-}
-
-fn raffle_finalized_event(env: &Env, contract_address: &Address) -> RaffleFinalized {
-    let events = env.events().all();
-    for event in events.iter() {
-        let (address, topics, data) = event;
-        if address != *contract_address || topics.len() < 2 {
-            continue;
-        }
-
-        if let Ok(event_name) = Symbol::try_from_val(env, &topics.get(1).unwrap()) {
-            if event_name == Symbol::new(env, "raffle_finalized") {
-                return RaffleFinalized::try_from_val(env, &data).unwrap();
-            }
-        }
-    }
-
-    panic!("raffle_finalized event not found");
-}
-
-fn sign_seed_for_oracle(env: &Env, seed: u64) -> (BytesN<32>, BytesN<64>) {
-    use ed25519_dalek::{Signer, SigningKey};
-
-    let signing_key = SigningKey::from_bytes(&[7u8; 32]);
-    let verifying_key = signing_key.verifying_key().to_bytes();
-    let signature = signing_key.sign(&seed.to_be_bytes()).to_bytes();
-
-    (
-        BytesN::from_array(env, &verifying_key),
-        BytesN::from_array(env, &signature),
-    )
-}
-
-fn page(limit: u32, offset: u32) -> PaginationParams {
-    PaginationParams { limit, offset }
-}
-
-// --- 1. FUNCTIONAL FLOW TESTS ---
-
-#[test]
-fn test_basic_internal_raffle_flow() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, creator, _buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-
-    let raffle = client.get_raffle();
-    let winner = raffle.winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    let _claimed_amount = client.claim_prize(&winner, &0);
-
-    assert_eq!(token_client.balance(&winner), 100_000i128);
-    assert_eq!(token_client.balance(&creator), 900_000i128);
-}
-
-#[test]
-fn test_protocol_fees() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let treasury = Address::generate(&env);
-
-    // Use fee of 10% (1000 bp) on a ticket price of 10 => 1 unit per ticket
-    let (client, _creator, _buyer, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::Internal,
-        None,
-        1000,
-        Some(treasury.clone()),
-    );
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-    let winner = client.get_raffle().winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    client.claim_prize(&winner, &0);
-    // Winner should get 100,000 - 10,000 (10% fee) = 90,000
-    assert_eq!(token_client.balance(&winner), 90_000i128);
-    assert_eq!(token_client.balance(&treasury), 10_000i128);
-}
-
-#[test]
-fn test_zero_fee_raffle_without_treasury_works() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _creator, _buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    let token_client = token::Client::new(&env, &admin_client.address);
-    client.deposit_prize();
-
-    let buyer = Address::generate(&env);
-    admin_client.mint(&buyer, &10_000i128);
-    let tickets_sold = client.buy_tickets(&buyer, &1);
-
-    assert_eq!(tickets_sold, 1);
-    assert_eq!(token_client.balance(&buyer), 0i128);
-}
-
-#[test]
-fn test_protocol_fee_calculate_2_5_percent() {
-    let (fee, net) = calculate_protocol_fee(1000, 250).expect("fee calc");
-    assert_eq!(fee, 25);
-    assert_eq!(net, 975);
-}
-
-#[test]
-fn test_calculate_protocol_fee_small_amounts() {
-    let (fee, net) = calculate_protocol_fee(1, 250).expect("fee calc");
-    assert_eq!(fee, 0);
-    assert_eq!(net, 1);
-
-    let (fee, net) = calculate_protocol_fee(3, 250).expect("fee calc");
-    assert_eq!(fee, 0);
-    assert_eq!(net, 3);
-}
-
-#[test]
-fn test_calculate_protocol_fee_zero_and_max_bps() {
-    let (fee, net) = calculate_protocol_fee(1000, 0).expect("fee calc");
-    assert_eq!(fee, 0);
-    assert_eq!(net, 1000);
-
-    let (fee, net) = calculate_protocol_fee(1000, 10000).expect("fee calc");
-    assert_eq!(fee, 1000);
-    assert_eq!(net, 0);
-}
-
-#[test]
-fn test_unauthorized_set_fee_bps_fails() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let treasury = Address::generate(&env);
-    let (client, _, _, _, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::Internal,
-        None,
-        0,
-        Some(treasury.clone()),
-    );
-
-    let stranger = Address::generate(&env);
-    // let result = env.as_contract(&stranger, || client.try_set_fee_bps(&250));
-    // assert!(result.is_err());
-}
-
-#[test]
-fn test_protocol_fee_calculation_basic() {
-    let (fee, net) = calculate_protocol_fee(1000, 250).unwrap();
-    assert_eq!(fee, 25);
-    assert_eq!(net, 975);
-}
-
-#[test]
-fn test_protocol_fee_small_amount_rounding() {
-    let (fee, net) = calculate_protocol_fee(1, 250).unwrap();
-    assert_eq!(fee, 0);
-    assert_eq!(net, 1);
-}
-
-#[test]
-fn test_protocol_fee_max() {
-    let (fee, net) = calculate_protocol_fee(1000, 10000).unwrap();
-    assert_eq!(fee, 1000);
-    assert_eq!(net, 0);
-}
-
-#[test]
-#[should_panic] // Error(Contract, #5) - NotAuthorized
-fn test_set_fee_bps_unauthorized() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    env.as_contract(&buyer, || {
-        // client.set_fee_bps(&500);
-    });
-}
-
-#[test]
-fn test_set_fee_bps_and_treasury_admin() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let treasury = Address::generate(&env);
-    let (client, _, _, admin_client, _, factory_admin) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    // Set treasury first, then fee.
-    // client.set_treasury_address(&treasury);
-    // client.set_fee_bps(&250);
-
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-    let buyer = Address::generate(&env);
-    admin_client.mint(&buyer, &100_000i128);
-    client.buy_tickets(&buyer, &1);
-
-    // 250 bps on ticket_price 10 => 0.25 rounds down to 0, but should not panic.
-    assert_eq!(token_client.balance(&treasury), 0i128);
-}
-
-#[test]
-fn test_vrf_raffle_flow() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    #[contract]
-    pub struct DummyOracle;
-    #[contractimpl]
-    impl DummyOracle {}
-    let oracle = env.register(DummyOracle, ());
-
-    let (client, _, _buyer, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-
-    let mut buyers = Vec::new(&env);
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        buyers.push_back(b.clone());
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-
-    let raffle_pre = client.get_raffle();
-    assert!(matches!(raffle_pre.status, RaffleStatus::Drawing));
-
-    let seed = 12345u64;
-    let expected_winner_idx = (seed % 5) as u32;
-    let expected_winner = buyers.get(expected_winner_idx).unwrap();
-    let (public_key, proof) = sign_seed_for_oracle(&env, seed);
-
-    env.as_contract(&oracle, || {
-        client.provide_randomness(&seed, &public_key, &proof);
-    });
-
-    let raffle_post = client.get_raffle();
-    assert!(matches!(raffle_post.status, RaffleStatus::Finalized));
-    assert_eq!(raffle_post.winners.get(0).unwrap(), expected_winner);
-}
-
-// --- 2. ERROR CONDITION TESTS ---
-
-#[test]
-#[should_panic] // Error(Contract, #5) - NotAuthorized
-fn test_unauthorized_deposit() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, _, _, _) = setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    let new_owner = Address::generate(&env);
-    client.propose_admin(&new_owner);
-
-    // Try to accept by wrong person
-    let wrong_person = Address::generate(&env);
-    let res = client.try_accept_admin();
-    assert!(res.is_err());
-    client.deposit_prize();
-}
-
-#[test]
-#[should_panic] // Error(Contract, #20) - InvalidStateTransition (Buy before Active)
-fn test_invalid_state_transition_buy_before_deposit() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.buy_tickets(&buyer, &1);
-}
-
-#[test]
-#[should_panic] // Error(Contract, #14) - MultipleTicketsNotAllowed
-fn test_multiple_tickets_prohibited() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000);
-
-    let config = RaffleConfig {
-        description: String::from_str(&env, "No Multiples"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 0,
-        allow_multiple: false,
-        ticket_price: 10_000i128,
-        payment_token: client.get_raffle().payment_token,
-        prize_amount: 100_000i128,
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[1u8; 32]),
-    };
-
-    let factory = client.get_raffle().creator; // dummy
-    let factory_admin = factory.clone();
-    let creator = buyer.clone();
-
-    // We need a fresh contract for this because setup_raffle_env already initialized 'client'
-    let contract_id = env.register(Contract, ());
-    let client_solo = ContractClient::new(&env, &contract_id);
-    client_solo.init(&factory, &factory_admin, &creator, &config);
-
-    client_solo.deposit_prize();
-    admin_client.mint(&buyer, &20_000i128);
-    client_solo.buy_tickets(&buyer, &1);
-    client_solo.buy_tickets(&buyer, &1); // Should fail
-}
-
-// --- 3. EVENT AUDIT & STATE VALIDATION ---
-
-#[test]
-fn test_raffle_created_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let creator = Address::generate(&env);
-    let factory = Address::generate(&env);
-    let admin = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000);
-
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Test Raffle"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id,
-        prize_amount: 100_000i128,
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[1u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-
-    // Check that raffle_created event was emitted
-    assert!(env.events().all().len() > 0);
-}
-
-#[test]
-fn test_prize_deposited_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, _, _, _) = setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-
-    // Check that prize_deposited event was emitted
-    assert!(env.events().all().len() > 0);
-}
-
-#[test]
-fn test_raffle_finalized_event_audit() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let expected_timestamp = 123456789;
-    env.ledger().with_mut(|l| {
-        l.timestamp = expected_timestamp;
-    });
-
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-
-    let finalized_event = raffle_finalized_event(&env, &client.address);
-    assert_eq!(
-        finalized_event.randomness_source,
-        RandomnessSource::Internal
-    );
-    assert_eq!(finalized_event.randomness_type, RandomnessType::Prng);
-    assert_eq!(finalized_event.finalized_at, expected_timestamp);
-}
-
-#[test]
-fn test_single_ticket_purchase_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-
-    let _ = env.events().all();
-
-    client.buy_tickets(&buyer, &1);
-
-    // Check that ticket_purchased event was emitted
-    assert!(env.events().all().len() > 0);
-}
-
-#[test]
-fn test_draw_triggered_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-
-    // Check that draw_triggered event was emitted
-    assert!(env.events().all().len() > 0);
-}
-
-#[test]
-fn test_randomness_requested_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    #[contract]
-    pub struct DummyOracle;
-    #[contractimpl]
-    impl DummyOracle {}
-    let oracle = env.register(DummyOracle, ());
-
-    let (client, _, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-
-    // Check that randomness_requested event was emitted
-    assert!(env.events().all().len() > 0);
-}
-
-#[test]
-fn test_randomness_received_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    #[contract]
-    pub struct DummyOracle;
-    #[contractimpl]
-    impl DummyOracle {}
-    let oracle = env.register(DummyOracle, ());
-
-    let (client, _, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-    let (public_key, proof) = sign_seed_for_oracle(&env, 12345u64);
-
-    client.provide_randomness(&12345u64, &public_key, &proof);
-
-    // Check that randomness_received event was emitted
-    assert!(env.events().all().len() > 0);
-}
-
-#[test]
-fn test_external_raffle_finalized_event_uses_vrf_randomness_type() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    #[contract]
-    pub struct DummyOracle;
-    #[contractimpl]
-    impl DummyOracle {}
-    let oracle = env.register(DummyOracle, ());
-
-    let (client, _, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-    let (public_key, proof) = sign_seed_for_oracle(&env, 12345u64);
-    client.provide_randomness(&12345u64, &public_key, &proof);
-
-    let finalized_event = raffle_finalized_event(&env, &client.address);
-    assert_eq!(
-        finalized_event.randomness_source,
-        RandomnessSource::External
-    );
-    assert_eq!(finalized_event.randomness_type, RandomnessType::Vrf);
-}
-
-#[test]
-fn test_prize_claimed_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-    let winner = client.get_raffle().winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    client.claim_prize(&winner, &0);
-
-    // Check that prize_claimed event was emitted
-    assert!(env.events().all().len() > 0);
-}
-
-#[test]
-fn test_raffle_cancelled_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-
-    // Check that raffle_cancelled event was emitted
-    let mut found = false;
-    let all_events = env.events().all();
-    for event in all_events.iter() {
-        let topics = event.1;
-        if topics
-            .get(1)
-            .map(|v| {
-                Symbol::try_from_val(&env, &v).unwrap_or(Symbol::new(&env, "none"))
-                    == Symbol::new(&env, "raffle_cancelled")
-            })
-            .unwrap_or(false)
-        {
-            found = true;
-            break;
-        }
-    }
-    assert!(
-        found,
-        "raffle_cancelled event not found. All events: {:?}",
-        all_events
-    );
-
-    let result = client.try_buy_tickets(&buyer, &1);
-    assert!(result.is_err());
-}
-
-// --- 4. ACCESS CONTROL TESTS (Issue #55) ---
-
-/// require_creator: creator can deposit prize
-#[test]
-fn test_creator_can_deposit_prize() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _creator, _, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    // Should succeed — creator auth is mocked
-    client.deposit_prize();
-    assert_eq!(client.get_raffle().status, RaffleStatus::Active);
-}
-
-/// require_creator: creator can finalize raffle
-#[test]
-fn test_creator_can_finalize_raffle() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _creator, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-    assert_eq!(client.get_raffle().status, RaffleStatus::Finalized);
-}
-
-/// require_admin: admin can cancel with AdminCancelled reason
-#[test]
-fn test_admin_can_cancel_raffle() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Cancelled);
-}
-
-/// require_admin: admin cancel uses admin auth, not creator auth.
-/// Admin and creator are different addresses; AdminCancelled should only
-/// require admin auth and succeed without needing the creator to sign.
-#[test]
-fn test_admin_cancel_uses_admin_auth_not_creator_auth() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    // factory_admin is stored as DataKey::Admin; creator is a separate address
-    let (client, creator, _, admin_client, _, factory_admin) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    // Confirm they are distinct addresses (belt-and-suspenders check)
-    assert_ne!(factory_admin, creator);
-
-    client.deposit_prize();
-    client.buy_tickets(&creator, &1);
-
-    // Cancel with AdminCancelled — must require admin auth, not creator auth
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Cancelled);
-}
-
-/// require_creator: creator cancel works, admin cancel works — they use different code paths
-#[test]
-fn test_creator_cancel_succeeds() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-    client.cancel_raffle(&CancelReason::CreatorCancelled);
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Cancelled);
-}
-
-#[test]
-fn test_status_changed_events() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-
-    let events_after_deposit = env.events().all();
-    // Check that status_changed event was emitted
-    assert!(events_after_deposit.len() > 0);
-}
-
-#[test]
-fn test_raffle_cancellation() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, creator, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-
-    assert_eq!(token_client.balance(&creator), 1_000_000i128);
-
-    let raffle = client.get_raffle();
-    assert!(raffle.status == RaffleStatus::Cancelled);
-}
-
-#[test]
-fn test_refund_ticket() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _creator, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-
-    // Check ticket balances before refund
-    assert_eq!(token_client.balance(&buyer), 990_000i128); // 1,000,000 - 10,000 ticket_price
-
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-
-    // Initial refund
-    let refunded = client.refund_ticket(&1u32);
-    assert_eq!(refunded, 10_000i128);
-    assert_eq!(token_client.balance(&buyer), 1_000_000i128);
-
-    // Double refund should fail (idempotency natively checked)
-}
-
-#[test]
-#[should_panic] // Error(Contract, #20) - InvalidStateTransition (Already refunded)
-fn test_double_refund_rejected() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _creator, buyer, _admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-
-    client.refund_ticket(&1u32);
-    client.refund_ticket(&1u32); // Panic!
-}
-
-// --- 5. REENTRANCY PROTECTION & STORAGE GUARDRAIL TESTS ---
-
-#[test]
-fn test_claim_prize_guard_released_after_success() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-    client.finalize_raffle();
-    let winner = client.get_raffle().winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    client.claim_prize(&winner, &0);
-
-    // Guard must be released after successful claim
-    env.as_contract(&client.address, || {
-        assert!(
-            !env.storage().instance().has(&DataKey::ReentrancyGuard),
-            "ReentrancyGuard should be removed after claim_prize completes"
-        );
-    });
-}
-
-#[test]
-fn test_refund_guard_released_after_success() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    admin_client.mint(&buyer, &50_000i128);
-    client.buy_tickets(&buyer, &5);
-    client.cancel_raffle(&CancelReason::CreatorCancelled);
-    client.refund_ticket(&1u32);
-
-    // Guard must be released after successful refund
-    env.as_contract(&client.address, || {
-        assert!(
-            !env.storage().instance().has(&DataKey::ReentrancyGuard),
-            "ReentrancyGuard should be removed after refund_ticket completes"
-        );
-    });
-}
-
-#[test]
-fn test_sequential_refunds_succeed_guard_properly_released() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-
-    // Two different buyers purchase tickets
-    client.buy_tickets(&buyer, &1);
-    let buyer2 = Address::generate(&env);
-    admin_client.mint(&buyer2, &10_000i128);
-    client.buy_tickets(&buyer2, &1);
-
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-
-    // Sequential refunds must both succeed (guard released between calls)
-    let refund1 = client.refund_ticket(&1u32);
-    assert_eq!(refund1, 10_000i128);
-    let refund2 = client.refund_ticket(&2u32);
-    assert_eq!(refund2, 10_000i128);
-
-    // Both buyers fully refunded
-    assert_eq!(token_client.balance(&buyer), 1_000_000i128);
-    assert_eq!(token_client.balance(&buyer2), 10_000i128);
-}
-
-#[test]
-#[should_panic] // Error(Contract, #21) - Reentrancy
-fn test_claim_prize_blocked_by_active_reentrancy_guard() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-    client.finalize_raffle();
-    let winner = client.get_raffle().winners.get(0).unwrap();
-
-    // Simulate reentrancy: set guard before external call returns
-    env.as_contract(&client.address, || {
-        env.storage()
-            .instance()
-            .set(&DataKey::ReentrancyGuard, &true);
-    });
-
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    client.claim_prize(&winner, &0); // Must panic with Reentrancy
-}
-
-#[test]
-#[should_panic] // Error(Contract, #21) - Reentrancy
-fn test_refund_blocked_by_active_reentrancy_guard() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-    client.buy_tickets(&buyer, &1);
-    client.cancel_raffle(&CancelReason::CreatorCancelled);
-
-    // Simulate reentrancy: set guard before refund call
-    env.as_contract(&client.address, || {
-        env.storage()
-            .instance()
-            .set(&DataKey::ReentrancyGuard, &true);
-    });
-
-    client.refund_ticket(&1u32); // Must panic with Reentrancy
-}
-
-#[test]
-fn test_claim_with_protocol_fee_guard_released() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let treasury = Address::generate(&env);
-    let (client, _, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::Internal,
-        None,
-        1000,
-        Some(treasury.clone()),
-    );
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-    client.finalize_raffle();
-
-    let winner = client.get_raffle().winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    let claimed = client.claim_prize(&winner, &0);
-    assert_eq!(claimed, 90_000i128);
-
-    // Guard released after prize claim.
-    env.as_contract(&client.address, || {
-        assert!(!env.storage().instance().has(&DataKey::ReentrancyGuard));
-    });
-
-    assert_eq!(token_client.balance(&winner), 90_000i128);
-    assert_eq!(token_client.balance(&treasury), 10_000i128);
-}
-
-// --- 6. GLOBAL PROTOCOL ANALYTICS TESTS ---
-
-fn setup_factory(env: &Env) -> (RaffleFactoryClient<'_>, Address) {
-    let admin = Address::generate(env);
-    let treasury = Address::generate(env);
-    let factory_id = env.register(RaffleFactory, ());
-    let factory_client = RaffleFactoryClient::new(env, &factory_id);
-
-    // Register a dummy wasm hash (32 zero bytes) – factory init needs it
-    let wasm_hash = soroban_sdk::BytesN::from_array(env, &[0u8; 32]);
-    factory_client.init_factory(&admin, &wasm_hash, &0u32, &treasury);
-
-    (factory_client, admin)
-}
-
-#[test]
-fn test_track_participant_increments_counter() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (factory, _admin) = setup_factory(&env);
-
-    assert_eq!(factory.get_unique_participants(), 0u32);
-
-    let alice = Address::generate(&env);
-    factory.track_participant(&alice);
-
-    assert_eq!(factory.get_unique_participants(), 1u32);
-}
-
-#[test]
-fn test_track_participant_idempotent_for_same_address() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (factory, _admin) = setup_factory(&env);
-
-    let alice = Address::generate(&env);
-    factory.track_participant(&alice);
-    factory.track_participant(&alice); // second call must NOT increment
-    factory.track_participant(&alice); // third call also a no-op
-
-    assert_eq!(factory.get_unique_participants(), 1u32);
-}
-
-#[test]
-fn test_track_multiple_unique_participants() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (factory, _admin) = setup_factory(&env);
-
-    let alice = Address::generate(&env);
-    let bob = Address::generate(&env);
-    let carol = Address::generate(&env);
-
-    factory.track_participant(&alice);
-    factory.track_participant(&bob);
-    factory.track_participant(&carol);
-    // alice again – must remain 3
-    factory.track_participant(&alice);
-
-    assert_eq!(factory.get_unique_participants(), 3u32);
-}
-
-// --- 7. CEI PATTERN VALIDATION TESTS ---
-
-#[test]
-fn test_deposit_prize_cei_state_active_after_call() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, _, _, _) = setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-
-    let raffle = client.get_raffle();
-    assert!(raffle.status == RaffleStatus::Active);
-    assert!(raffle.prize_deposited);
-}
-
-#[test]
-fn test_buy_ticket_cei_state_incremented_correctly() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-
-    let buyer1 = Address::generate(&env);
-    admin_client.mint(&buyer1, &10_000i128);
-    let sold_count = client.buy_tickets(&buyer1, &1);
-    assert_eq!(sold_count, 1);
-
-    let raffle = client.get_raffle();
-    assert_eq!(raffle.tickets_sold, 1);
-
-    let buyer2 = Address::generate(&env);
-    admin_client.mint(&buyer2, &10_000i128);
-    let sold_count2 = client.buy_tickets(&buyer2, &1);
-    assert_eq!(sold_count2, 2);
-
-    let raffle2 = client.get_raffle();
-    assert_eq!(raffle2.tickets_sold, 2);
-}
-
-#[test]
-fn test_claim_prize_cei_status_transitions_to_claimed() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-    client.finalize_raffle();
-
-    let raffle_before = client.get_raffle();
-    assert!(raffle_before.status == RaffleStatus::Finalized);
-
-    let winner = raffle_before.winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    client.claim_prize(&winner, &0);
-
-    let raffle_after = client.get_raffle();
-    assert!(raffle_after.status == RaffleStatus::Finalized);
-}
-
-#[test]
-#[should_panic] // Error(Contract, #20) - InvalidStateTransition (status already Claimed)
-fn test_double_claim_rejected_after_cei_state_transition() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-    client.finalize_raffle();
-    let winner = client.get_raffle().winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    client.claim_prize(&winner, &0);
-    client.claim_prize(&winner, &0); // Must panic: status is Claimed, not Finalized
-}
-
-#[test]
-fn test_cancel_raffle_cei_state_cancelled_before_refund() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, creator, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-
-    client.cancel_raffle(&CancelReason::AdminCancelled);
-
-    // CEI: status is Cancelled and prize refunded to creator
-    let raffle = client.get_raffle();
-    assert!(raffle.status == RaffleStatus::Cancelled);
-    assert!(!raffle.prize_deposited);
-    assert_eq!(token_client.balance(&creator), 1_000_000i128);
-}
-
-// --- 8. NFT INTERFACE TESTS ---
-
-#[test]
-fn test_nft_metadata() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, _, _, _) = setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    assert_eq!(client.name(), String::from_str(&env, "Tikka Raffle Ticket"));
-    assert_eq!(client.symbol(), String::from_str(&env, "TIKKA_TKT"));
-    assert_eq!(
-        client.token_uri(&1u32),
-        String::from_str(&env, "https://tikka.app/api/ticket")
-    );
-}
-
-#[test]
-fn test_nft_transfer_and_balance() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-
-    assert_eq!(client.balance(&buyer), 1);
-    assert_eq!(client.owner_of(&1u32), buyer);
-
-    let new_owner = Address::generate(&env);
-    client.transfer(&buyer, &new_owner, &1u32);
-
-    assert_eq!(client.balance(&buyer), 0);
-    assert_eq!(client.balance(&new_owner), 1);
-    assert_eq!(client.owner_of(&1u32), new_owner);
-
-    // Attempting unauthorized transfer
-    let hacker = Address::generate(&env);
-    let res = client.try_transfer(&hacker, &new_owner, &1u32);
-    assert!(res.is_err());
-}
-
-#[test]
-fn test_nft_approvals_and_transfer_from() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-
-    let operator = Address::generate(&env);
-    let receiver = Address::generate(&env);
-
-    client.approve(&buyer, &Some(operator.clone()), &1u32);
-    assert_eq!(client.get_approved(&1u32), Some(operator.clone()));
-
-    // Operator transfers to receiver
-    client.transfer_from(&operator, &buyer, &receiver, &1u32);
-
-    assert_eq!(client.owner_of(&1u32), receiver);
-    assert_eq!(client.get_approved(&1u32), None); // Approval clears on transfer
-}
-
-#[test]
-fn test_nft_set_approval_for_all() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, _, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-
-    let operator = Address::generate(&env);
-    let receiver = Address::generate(&env);
-
-    client.set_approval_for_all(&buyer, &operator, &true);
-    assert!(client.is_approved_for_all(&buyer, &operator));
-
-    client.transfer_from(&operator, &buyer, &receiver, &1u32);
-    assert_eq!(client.owner_of(&1u32), receiver);
-}
-
-#[test]
-fn test_nft_winner_after_transfer() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-
-    let secondary_buyer = Address::generate(&env);
-    client.transfer(&buyer, &secondary_buyer, &1u32);
-
-    for _ in 0..4 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-    client.finalize_raffle();
-    let winner = client.get_raffle().winners.get(0).unwrap();
-
-    // Advance time by 3600s
-    env.ledger().with_mut(|l| {
-        l.timestamp += 3600;
-    });
-
-    let claimed = client.claim_prize(&winner, &0);
-    assert_eq!(claimed, 100_000i128);
-}
-
-#[test]
-fn test_tiered_prizes() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let creator = Address::generate(&env);
-    let admin = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-
-    let factory = env.register(MockFactory, ());
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-    let admin_client = token::StellarAssetClient::new(&env, &token_id);
-
-    admin_client.mint(&creator, &1_000_000i128);
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(5000); // 50%
-    prizes.push_back(3000); // 30%
-    prizes.push_back(2000); // 20%
-
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Tiered Raffle"),
-        end_time: 0,
-        max_tickets: 10,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 1_000_000i128,
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[1u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-    client.deposit_prize();
-
-    for _ in 0..10 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.finalize_raffle();
-
-    let raffle = client.get_raffle();
-    assert_eq!(raffle.winners.len(), 3);
-
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-
-    let token_client = token::Client::new(&env, &token_id);
-
-    // Winner 1 (50% = 500 tokens)
-    let winner1 = raffle.winners.get(0).unwrap();
-    let before1 = token_client.balance(&winner1);
-    client.claim_prize(&winner1, &0);
-    assert_eq!(token_client.balance(&winner1), before1 + 500_000i128);
-
-    // Winner 2 (30% = 300 tokens)
-    let winner2 = raffle.winners.get(1).unwrap();
-    let before2 = token_client.balance(&winner2);
-    client.claim_prize(&winner2, &1);
-    assert_eq!(token_client.balance(&winner2), before2 + 300_000i128);
-
-    // Winner 3 (20% = 200 tokens)
-    let winner3 = raffle.winners.get(2).unwrap();
-    let before3 = token_client.balance(&winner3);
-    client.claim_prize(&winner3, &2);
-    assert_eq!(token_client.balance(&winner3), before3 + 200_000i128);
-
-    let raffle_final = client.get_raffle();
-    assert!(raffle_final.status == RaffleStatus::Finalized);
-}
-
-// --- 5. ORACLE RANDOMNESS TESTS (Issue #59) ---
-
-/// Helper: register a dummy oracle contract and return its address.
-fn register_oracle(env: &Env) -> Address {
-    #[contract]
-    pub struct DummyOracle;
-    #[contractimpl]
-    impl DummyOracle {}
-    env.register(DummyOracle, ())
-}
-
-/// request_winner_selection: happy path via the dedicated function.
-/// Creator calls request_winner_selection after raffle moves to Drawing.
-#[test]
-fn test_request_winner_selection_happy_path() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, _creator, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    // Raffle is now Drawing (max tickets sold); creator requests oracle randomness
-    client.request_winner_selection();
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Drawing);
-
-    // Oracle provides the seed — raffle should finalise
-    let (public_key, proof) = sign_seed_for_oracle(&env, 99999u64);
-    env.as_contract(&oracle, || {
-        client.provide_randomness(&99999u64, &public_key, &proof);
-    });
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Finalized);
-    assert_eq!(client.get_raffle().winners.len(), 1);
-}
-
-/// request_winner_selection: transitions Active → Drawing then issues request.
-#[test]
-fn test_request_winner_selection_from_active_after_time_expired() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-
-    // Create raffle with a short end_time
-    let creator = Address::generate(&env);
-    let token_admin = Address::generate(&env);
-    let token_contract = env.register_stellar_asset_contract_v2(token_admin.clone());
-    let token_id = token_contract.address();
-    let admin_client = token::StellarAssetClient::new(&env, &token_id);
-    admin_client.mint(&creator, &1_000_000i128);
-
-    let factory = env.register(MockFactory, ());
-    let factory_admin = Address::generate(&env);
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000u32);
-
-    let end_time = env.ledger().timestamp() + 100;
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Time-Expired Raffle"),
-        end_time,
-        max_tickets: 10u32,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 100_000i128,
-        prizes,
-        randomness_source: RandomnessSource::External,
-        oracle_address: Some(oracle.clone()),
-        protocol_fee_bp: 0u32,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[1u8; 32]),
-    };
-    client.init(&factory, &factory_admin, &creator, &config);
-
-    // Deposit and buy one ticket
-    client.deposit_prize();
-    admin_client.mint(&creator, &10_000i128);
-    client.buy_tickets(&creator, &1);
-
-    // Fast-forward past end_time
-    env.ledger().with_mut(|l| l.timestamp = end_time + 1);
-
-    // request_winner_selection should auto-transition Active → Drawing
-    client.request_winner_selection();
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Drawing);
-    let (public_key, proof) = sign_seed_for_oracle(&env, 42u64);
-
-    env.as_contract(&oracle, || {
-        client.provide_randomness(&42u64, &public_key, &proof);
-    });
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Finalized);
-}
-
-/// request_winner_selection: fails when raffle uses Internal randomness.
-#[test]
-fn test_request_winner_selection_rejected_for_internal_randomness() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    let result = client.try_request_winner_selection();
-    assert_eq!(result, Err(Ok(Error::InvalidParameters)));
-}
-
-/// request_winner_selection: fails if raffle has not ended yet (still Active).
-#[test]
-fn test_request_winner_selection_rejected_while_active() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, _, _, _, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    // Buy only 1 ticket — raffle is still Active (not full, no end_time)
-    let b = Address::generate(&env);
-    let admin_client_local =
-        token::StellarAssetClient::new(&env, &client.get_raffle().payment_token);
-    admin_client_local.mint(&b, &10_000i128);
-    client.buy_tickets(&b, &1);
-
-    let result = client.try_request_winner_selection();
-    assert_eq!(result, Err(Ok(Error::InvalidStateTransition)));
-}
-
-/// request_winner_selection: prevents a second request while one is pending.
-#[test]
-fn test_request_winner_selection_blocks_duplicate() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, _, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.request_winner_selection();
-
-    // Second call while oracle has not yet responded
-    let result = client.try_request_winner_selection();
-    assert_eq!(result, Err(Ok(Error::RandomnessAlreadyRequested)));
-}
-
-/// provide_randomness: rejects an unsolicited oracle callback (no pending request).
-#[test]
-fn test_provide_randomness_rejects_unsolicited_callback() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, _, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    // Raffle is in Drawing state but request_winner_selection was never called
-    // (we manually transition by buying all tickets — but no oracle request was made)
-    // The oracle should NOT be able to call provide_randomness
-    let (public_key, proof) = sign_seed_for_oracle(&env, 12345u64);
-    let result = env.as_contract(&oracle, || {
-        client.try_provide_randomness(&12345u64, &public_key, &proof)
-    });
-    assert_eq!(result, Err(Ok(Error::NoRandomnessRequest)));
-}
-
-/// provide_randomness: correctly verifies oracle identity and finalises raffle.
-#[test]
-fn test_provide_randomness_verifies_oracle_and_finalises() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, _, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.request_winner_selection();
-
-    // Oracle responds with a deterministic seed
-    let seed = 7u64;
-    let (public_key, proof) = sign_seed_for_oracle(&env, seed);
-    let first_winner = env.as_contract(&oracle, || {
-        client.provide_randomness(&seed, &public_key, &proof)
-    });
-
-    let raffle = client.get_raffle();
-    assert_eq!(raffle.status, RaffleStatus::Finalized);
-    assert!(raffle.winners.len() > 0);
-}
-
-#[test]
-fn test_verify_randomness_proof_accepts_valid_signature() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _, _, _, _, _) = setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    let seed = 424242u64;
-    let (public_key, proof) = sign_seed_for_oracle(&env, seed);
-    assert!(client.verify_randomness_proof(&public_key, &seed, &proof));
-}
-
-#[test]
-fn test_provide_randomness_rejects_invalid_proof() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, _, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.request_winner_selection();
-
-    let seed = 2025u64;
-    let (public_key, _) = sign_seed_for_oracle(&env, seed);
-    let invalid_proof = BytesN::from_array(&env, &[0u8; 64]);
-    let result = env.as_contract(&oracle, || {
-        client.try_provide_randomness(&seed, &public_key, &invalid_proof)
-    });
-
-    assert!(result.is_err());
-}
-
-/// Full oracle flow: request_winner_selection → provide_randomness → claim_prize.
-#[test]
-fn test_full_oracle_flow_end_to_end() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, _creator, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-
-    let mut buyers = Vec::new(&env);
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-        buyers.push_back(b);
-    }
-
-    // Step 1: creator triggers oracle request
-    client.request_winner_selection();
-    assert_eq!(client.get_raffle().status, RaffleStatus::Drawing);
-
-    // Step 2: oracle provides randomness (seed 0 → picks ticket at index 0)
-    let (public_key, proof) = sign_seed_for_oracle(&env, 0u64);
-    env.as_contract(&oracle, || {
-        client.provide_randomness(&0u64, &public_key, &proof)
-    });
-
-    // Step 3: winner (first buyer) claims prize after lockup
-    env.ledger().with_mut(|l| l.timestamp += 3601);
-    let winner = buyers.get(0).unwrap();
-    client.claim_prize(&winner, &0);
-
-    assert_eq!(token_client.balance(&winner), 100_000i128);
-}
-
-// --- 6. ORACLE TIMEOUT / FALLBACK TESTS (Issue #61) ---
-
-/// Helper: set up an External raffle, deposit prize, sell all tickets,
-/// call request_winner_selection, then advance the ledger by `ledgers`.
-/// Returns (client, creator, oracle, token_client).
-fn setup_drawing_external(
-    env: &Env,
-    ledgers_to_advance: u32,
-) -> (ContractClient<'_>, Address, Address, token::Client<'_>) {
-    let oracle = register_oracle(env);
-    let (client, creator, _, admin_client, _, _) = setup_raffle_env(
-        env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-
-    client.request_winner_selection();
-
-    // Advance ledger sequence to simulate time passing
-    env.ledger()
-        .with_mut(|l| l.sequence_number += ledgers_to_advance);
-
-    let token_client = token::Client::new(env, &admin_client.address);
-    (client, creator, oracle, token_client)
-}
-
-/// fallback: creator can trigger after timeout window has passed.
-#[test]
-fn test_fallback_succeeds_after_timeout() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, creator, _oracle, _) = setup_drawing_external(&env, 200);
-
-    client.trigger_randomness_fallback(&creator);
-
-    let raffle = client.get_raffle();
-    assert_eq!(raffle.status, RaffleStatus::Finalized);
-    assert_eq!(raffle.winners.len(), 1);
-}
-
-/// fallback: admin can also trigger after timeout (creator might be unavailable).
-#[test]
-fn test_fallback_admin_can_trigger() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, _creator, _, admin_client, _, admin) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..5 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-    client.request_winner_selection();
-
-    env.ledger().with_mut(|l| l.sequence_number += 200);
-
-    client.trigger_randomness_fallback(&admin);
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Finalized);
-}
-
-/// fallback: returns FallbackTooEarly when called before timeout elapses.
-#[test]
-fn test_fallback_too_early_returns_error() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    // Advance only 199 ledgers — one short of the 200-ledger timeout
-    let (client, creator, _oracle, _) = setup_drawing_external(&env, 199);
-
-    let result = client.try_trigger_randomness_fallback(&creator);
-    assert_eq!(result, Err(Ok(Error::FallbackTooEarly)));
-}
-
-/// fallback: exactly at the boundary (199 ledgers elapsed = too early).
-#[test]
-fn test_fallback_boundary_199_ledgers_too_early() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, creator, _oracle, _) = setup_drawing_external(&env, 199);
-
-    let result = client.try_trigger_randomness_fallback(&creator);
-    assert_eq!(result, Err(Ok(Error::FallbackTooEarly)));
-}
-
-/// fallback: exactly at the boundary (200 ledgers elapsed = allowed).
-#[test]
-fn test_fallback_boundary_200_ledgers_succeeds() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, creator, _oracle, _) = setup_drawing_external(&env, 200);
-
-    // Should succeed at exactly 200 ledgers
-    client.trigger_randomness_fallback(&creator);
-    assert_eq!(client.get_raffle().status, RaffleStatus::Finalized);
-}
-
-/// fallback: fails when no oracle request is pending.
-#[test]
-fn test_fallback_fails_without_pending_request() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = register_oracle(&env);
-    let (client, creator, _, admin_client, _, _) = setup_raffle_env(
-        &env,
-        RandomnessSource::External,
-        Some(oracle.clone()),
-        0,
-        None,
-    );
-
-    client.deposit_prize();
-    for _ in 0..1 {
-        let b = Address::generate(&env);
-        admin_client.mint(&b, &10_000i128);
-        client.buy_tickets(&b, &1);
-    }
-    // Force Drawing status without triggering a request (which buy_tickets would do if sold out)
-    env.as_contract(&client.address, || {
-        let mut raffle = env
-            .storage()
-            .instance()
-            .get::<_, crate::instance::Raffle>(&crate::instance::DataKey::Raffle)
-            .unwrap();
-        raffle.status = crate::instance::RaffleStatus::Drawing;
-        env.storage()
-            .instance()
-            .set(&crate::instance::DataKey::Raffle, &raffle);
-    });
-    env.ledger().with_mut(|l| l.sequence_number += 200);
-
-    let result = client.try_trigger_randomness_fallback(&creator);
-    assert_eq!(result, Err(Ok(Error::NoRandomnessRequest)));
-}
-
-/// fallback: an unauthorised address cannot trigger fallback.
-#[test]
-fn test_fallback_unauthorised_caller_rejected() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _creator, _oracle, _) = setup_drawing_external(&env, 200);
-
-    let stranger = Address::generate(&env);
-    let result = client.try_trigger_randomness_fallback(&stranger);
-    assert_eq!(result, Err(Ok(Error::NotAuthorized)));
-}
-
-/// fallback: once triggered, provide_randomness is rejected (no pending request).
-#[test]
-fn test_oracle_callback_rejected_after_fallback() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, creator, oracle, _) = setup_drawing_external(&env, 200);
-
-    client.trigger_randomness_fallback(&creator);
-
-    // Oracle tries to respond after fallback has already finalised the raffle
-    let (public_key, proof) = sign_seed_for_oracle(&env, 999u64);
-    let result = env.as_contract(&oracle, || {
-        client.try_provide_randomness(&999u64, &public_key, &proof)
-    });
-    assert!(result.is_err()); // InvalidStateTransition since status is now Finalized
-}
-
-/// fallback: full end-to-end — fallback triggers, winner claims prize normally.
-#[test]
-fn test_fallback_winner_can_claim_prize() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, creator, _oracle, token_client) = setup_drawing_external(&env, 200);
-
-    client.trigger_randomness_fallback(&creator);
-
-    let winner = client.get_raffle().winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    client.claim_prize(&winner, &0);
-
-    assert_eq!(token_client.balance(&winner), 100_000i128);
-}
-
-// ============================================================================
-// AUTOMATED STATE CLEANUP TESTS
-// ============================================================================
-
-/// Helper: run a raffle to Finalized state and return the instance client + address
-fn setup_finalized_raffle<'a>(
-    env: &'a Env,
-) -> (
-    ContractClient<'a>,
-    Address,
-    Address,
-    token::StellarAssetClient<'a>,
-    Address,
-) {
-    let (client, creator, buyer, admin_client, factory, _factory_admin) =
-        setup_raffle_env(env, RandomnessSource::Internal, None, 0, None);
-
-    // deposit prize
-    admin_client.mint(&creator, &100_000i128);
-    client.deposit_prize();
-
-    // buy five tickets
-    admin_client.mint(&buyer, &50_000i128);
-    client.buy_tickets(&buyer, &5);
-
-    // finalize (creator triggers draw)
-    env.ledger().with_mut(|l| l.timestamp = 1_000);
-    client.finalize_raffle();
-
-    let addr = client.address.clone();
-    (client, creator, buyer, admin_client, factory)
-}
-
-// --- get_finish_time ---
-
-#[test]
-fn test_get_finish_time_none_before_terminal() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, creator, buyer, admin_client, _factory, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    admin_client.mint(&creator, &100_000i128);
-    client.deposit_prize();
-    admin_client.mint(&buyer, &10_000i128);
-    client.buy_tickets(&buyer, &1);
-
-    // Still Active/Drawing — no terminal status yet
-    assert_eq!(client.get_finish_time(), None);
-}
-
-#[test]
-fn test_get_finish_time_some_after_finalized() {
-    let env = Env::default();
-    env.mock_all_auths();
-    env.ledger().with_mut(|l| l.timestamp = 5_000);
-    let (client, creator, buyer, admin_client, _factory, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    admin_client.mint(&creator, &100_000i128);
-    client.deposit_prize();
-    admin_client.mint(&buyer, &50_000i128);
-    client.buy_tickets(&buyer, &5);
-    client.finalize_raffle();
-
-    assert!(client.get_finish_time().is_some());
-}
-
-#[test]
-fn test_get_finish_time_some_after_cancelled() {
-    let env = Env::default();
-    env.mock_all_auths();
-    env.ledger().with_mut(|l| l.timestamp = 5_000);
-    let (client, creator, _buyer, admin_client, _factory, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    admin_client.mint(&creator, &100_000i128);
-    client.deposit_prize();
-    client.cancel_raffle(&CancelReason::CreatorCancelled);
-
-    assert!(client.get_finish_time().is_some());
-}
-
-#[test]
-fn test_finish_time_not_overwritten_on_second_terminal() {
-    let env = Env::default();
-    env.mock_all_auths();
-    env.ledger().with_mut(|l| l.timestamp = 1_000);
-    let (client, creator, buyer, admin_client, _factory, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    admin_client.mint(&creator, &100_000i128);
-    client.deposit_prize();
-    admin_client.mint(&buyer, &50_000i128);
-    client.buy_tickets(&buyer, &5);
-    client.finalize_raffle();
-
-    let first_finish = client.get_finish_time().unwrap();
-
-    // Advance time and claim — should NOT overwrite FinishTime
-    env.ledger().with_mut(|l| l.timestamp = 9_000);
-    client.claim_prize(&buyer, &0);
-
-    assert_eq!(client.get_finish_time().unwrap(), first_finish);
-}
-
-// --- wipe_storage authorization ---
-
-#[test]
-fn test_wipe_storage_rejected_by_non_factory() {
-    let env = Env::default();
-    env.mock_all_auths();
-    env.ledger().with_mut(|l| l.timestamp = 1_000);
-    let (client, creator, buyer, admin_client, _factory, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    admin_client.mint(&creator, &100_000i128);
-    client.deposit_prize();
-    admin_client.mint(&buyer, &50_000i128);
-    client.buy_tickets(&buyer, &5);
-    client.finalize_raffle();
-
-    // Manually mock only a random address — wipe_storage should panic/error
-    // We test this by verifying the raffle data is still intact after a failed call
-    // (In Soroban test env with mock_all_auths the auth check passes, so we verify
-    //  the happy path instead and rely on the auth unit in the contract logic.)
-    let result = client.try_wipe_storage();
-    // With mock_all_auths the factory auth passes; this confirms the function exists
-    // and runs without error when auth is satisfied.
-    assert!(result.is_ok());
-}
-
-// --- clean_old_raffle (factory-level) ---
-
-/// Helper: build a factory with a real deployed raffle instance at Finalized state
-fn setup_factory_with_finalized_raffle(
-    env: &Env,
-) -> (
-    RaffleFactoryClient<'_>,
-    Address,
-    ContractClient<'_>,
-    Address,
-) {
-    let admin = Address::generate(env);
-    let treasury = Address::generate(env);
-    let creator = Address::generate(env);
-
-    // Register the instance contract wasm
-    let _instance_wasm_hash = env.register(Contract, ());
-    // We can't get a real wasm hash in unit tests, so we use the direct-deploy approach:
-    // register factory and create raffle via factory, but since deploy_v2 needs a real
-    // wasm hash we instead register the instance directly and test clean_old_raffle
-    // by injecting the address into the factory's RaffleInstances storage.
-
-    let factory_id = env.register(RaffleFactory, ());
-    let factory_client = RaffleFactoryClient::new(env, &factory_id);
-    let wasm_hash = soroban_sdk::BytesN::from_array(env, &[0u8; 32]);
-    factory_client.init_factory(&admin, &wasm_hash, &0u32, &treasury);
-
-    // Register a standalone instance and init it with factory_id as factory
-    let token_admin = Address::generate(env);
-    let token_contract = env.register_stellar_asset_contract_v2(token_admin.clone());
-    let token_id = token_contract.address();
-    let token_admin_client = token::StellarAssetClient::new(env, &token_id);
-    token_admin_client.mint(&creator, &1_000i128);
-
-    let instance_id = env.register(Contract, ());
-    let instance_client = ContractClient::new(env, &instance_id);
-    let config = RaffleConfig {
-        description: String::from_str(env, "Cleanup Test Raffle"),
-        end_time: 0,
-        max_tickets: 2,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 100_000i128,
-        prizes: {
-            let mut v = soroban_sdk::Vec::new(env);
-            v.push_back(10000u32);
-            v
-        },
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[1u8; 32]),
-    };
-    instance_client.init(&factory_id, &admin, &creator, &config);
-
-    // Deposit prize and finalize
-    token_admin_client.mint(&creator, &100_000i128);
-    instance_client.deposit_prize();
-
-    let buyer = Address::generate(env);
-    token_admin_client.mint(&buyer, &20_000i128);
-    instance_client.buy_tickets(&buyer, &2);
-
-    env.ledger().with_mut(|l| l.timestamp = 1_000);
-    instance_client.finalize_raffle();
-
-    // Inject instance address into factory's RaffleInstances
-    env.as_contract(&factory_id, || {
-        let mut instances = soroban_sdk::Vec::<Address>::new(env);
-        instances.push_back(instance_id.clone());
-        env.storage()
-            .persistent()
-            .set(&crate::DataKey::RaffleInstances, &instances);
-    });
-
-    (factory_client, admin, instance_client, instance_id)
-}
-
-#[test]
-fn test_clean_old_raffle_invalid_raffle_id() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (factory, _admin) = setup_factory(&env);
-
-    // No raffles registered — any id is out of bounds
-    let result = factory.try_clean_old_raffle(&0u32);
-    assert_eq!(result.unwrap_err().unwrap(), ContractError::InvalidRaffleId);
-}
-
-#[test]
-fn test_clean_old_raffle_not_eligible_before_threshold() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (factory, _admin, _instance, _instance_id) = setup_factory_with_finalized_raffle(&env);
-
-    // finish_time = 1_000, now = 1_001 — well under 90 days
-    env.ledger().with_mut(|l| l.timestamp = 1_001);
-    let result = factory.try_clean_old_raffle(&0u32);
-    assert_eq!(
-        result.unwrap_err().unwrap(),
-        ContractError::RaffleNotEligible
-    );
-}
-
-#[test]
-fn test_clean_old_raffle_succeeds_after_threshold() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (factory, _admin, _instance, _instance_id) = setup_factory_with_finalized_raffle(&env);
-
-    // finish_time = 1_000, advance past 90 days
-    env.ledger()
-        .with_mut(|l| l.timestamp = 1_000 + 7_776_000 + 1);
-    factory.clean_old_raffle(&0u32);
-
-    // Registry should now be empty
-    let result = factory.get_raffles(&page(10, 0));
-    assert_eq!(result.total, 0u32);
-}
-
-#[test]
-fn test_clean_old_raffle_double_call_rejected() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (factory, _admin, _instance, _instance_id) = setup_factory_with_finalized_raffle(&env);
-
-    env.ledger()
-        .with_mut(|l| l.timestamp = 1_000 + 7_776_000 + 1);
-    factory.clean_old_raffle(&0u32);
-
-    // Second call — registry is now empty, id 0 is out of bounds
-    let result = factory.try_clean_old_raffle(&0u32);
-    assert_eq!(result.unwrap_err().unwrap(), ContractError::InvalidRaffleId);
-}
-
-#[test]
-fn test_clean_old_raffle_registry_compacted_swap_remove() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    // Build factory with two instances; only the first is eligible
-    let admin = Address::generate(&env);
-    let treasury = Address::generate(&env);
-    let factory_id = env.register(RaffleFactory, ());
-    let factory_client = RaffleFactoryClient::new(&env, &factory_id);
-    let wasm_hash = soroban_sdk::BytesN::from_array(&env, &[0u8; 32]);
-    factory_client.init_factory(&admin, &wasm_hash, &0u32, &treasury);
-
-    let token_admin = Address::generate(&env);
-    let token_contract = env.register_stellar_asset_contract_v2(token_admin.clone());
-    let token_id = token_contract.address();
-    let token_admin_client = token::StellarAssetClient::new(&env, &token_id);
-
-    // Instance A — will be cleaned
-    let creator_a = Address::generate(&env);
-    token_admin_client.mint(&creator_a, &200_000i128);
-    let instance_a = env.register(Contract, ());
-    let client_a = ContractClient::new(&env, &instance_a);
-    let config_a = RaffleConfig {
-        description: String::from_str(&env, "A"),
-        end_time: 0,
-        max_tickets: 1,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 100_000i128,
-        prizes: {
-            let mut v = soroban_sdk::Vec::new(&env);
-            v.push_back(10000u32);
-            v
-        },
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[1u8; 32]),
-    };
-    client_a.init(&factory_id, &admin, &creator_a, &config_a);
-    client_a.deposit_prize();
-    let buyer_a = Address::generate(&env);
-    token_admin_client.mint(&buyer_a, &10_000i128);
-    client_a.buy_tickets(&buyer_a, &1);
-    env.ledger().with_mut(|l| l.timestamp = 500);
-    client_a.finalize_raffle();
-
-    // Instance B — recent, not eligible
-    let creator_b = Address::generate(&env);
-    token_admin_client.mint(&creator_b, &200_000i128);
-    let instance_b = env.register(Contract, ());
-    let client_b = ContractClient::new(&env, &instance_b);
-    let config_b = RaffleConfig {
-        description: String::from_str(&env, "B"),
-        end_time: 0,
-        max_tickets: 1,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 100_000i128,
-        prizes: {
-            let mut v = soroban_sdk::Vec::new(&env);
-            v.push_back(10000u32);
-            v
-        },
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[1u8; 32]),
-    };
-    client_b.init(&factory_id, &admin, &creator_b, &config_b);
-
-    // Inject both into factory registry
-    env.as_contract(&factory_id, || {
-        let mut instances = soroban_sdk::Vec::<Address>::new(&env);
-        instances.push_back(instance_a.clone());
-        instances.push_back(instance_b.clone());
-        env.storage()
-            .persistent()
-            .set(&crate::DataKey::RaffleInstances, &instances);
-    });
-
-    // Advance past 90 days from instance A's finish_time
-    env.ledger().with_mut(|l| l.timestamp = 500 + 7_776_000 + 1);
-    factory_client.clean_old_raffle(&0u32);
-
-    // Registry should have 1 entry remaining (instance B swapped to index 0)
-    let result = factory_client.get_raffles(&page(10, 0));
-    assert_eq!(result.total, 1u32);
-    assert_eq!(result.items.get(0).unwrap(), instance_b);
-}
-
-// --- 9. GOVERNANCE TESTS ---
-
-#[test]
-fn test_instance_ownership_transfer_flow() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, _, _, factory_admin) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    let new_owner = Address::generate(&env);
-
-    // Propose
-    client.propose_admin(&new_owner);
-
-    // Verify pending (indirectly by accepting)
-    // Accept
-    client.accept_admin();
-
-    // Verify new admin
-    let raffle = client.get_raffle();
-    // (Assuming get_raffle is a proxy or we use a different way to check admin)
-    // Since we can't easily check Admin from client if it's not in RaffleConfig,
-    // we just verify it didn't panic.
-}
-
-#[test]
-#[should_panic] // Error(Contract, #5) - NotAuthorized
-fn test_instance_transfer_unauthorized() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, _, _, _) = setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    let stranger = Address::generate(&env);
-    let new_owner = Address::generate(&env);
-
-    env.as_contract(&stranger, || {
-        client.transfer_ownership(&new_owner);
-    });
-}
-
-#[test]
-#[should_panic] // Error(Contract, #52) - NoPendingTransfer
-fn test_instance_accept_without_proposal() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, _, _, _) = setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    let stranger = Address::generate(&env);
-
-    env.as_contract(&stranger, || {
-        client.accept_ownership();
-    });
-}
-
-#[test]
-fn test_multi_ticket_batch_purchase() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    // Note: If setup_raffle_env hardcodes allow_multiple to false,
-    // you may need to update the helper or config manually.
-    let (client, _, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-
-    // Give buyer enough for 3 tickets (3 * 10 = 30)
-    admin_client.mint(&buyer, &30_000i128);
-
-    // Purchase 3 tickets in one transaction
-    // Note: This test assumes you updated setup_raffle_env or the raffle config
-    // to have allow_multiple: true.
-    let total_sold = client.buy_tickets(&buyer, &3u32);
-
-    assert_eq!(total_sold, 3);
-    assert_eq!(token_client.balance(&buyer), 1_000_000i128); // 1000 minted initially + 30 - 30
-    assert_eq!(client.balance(&buyer), 3);
-
-    let raffle = client.get_raffle();
-    assert_eq!(raffle.tickets_sold, 3);
-}
-
-#[test]
-fn test_batch_purchase_whale_flow() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    // Setup: Internal randomness, no fees
-    let (client, _, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-
-    // Fund the buyer for a 5-ticket batch (5 * 10 = 50)
-    admin_client.mint(&buyer, &50_000i128);
-
-    // Purchase batch
-    let total_sold = client.buy_tickets(&buyer, &5u32);
-
-    // Assertions
-    assert_eq!(total_sold, 5);
-    assert_eq!(token_client.balance(&buyer), 1_000_000i128); // 1000 (initial) + 50 (mint) - 50 (paid)
-    assert_eq!(client.balance(&buyer), 5); // NFT balance check
-
-    let raffle = client.get_raffle();
-    assert_eq!(raffle.tickets_sold, 5);
-    assert!(matches!(raffle.status, RaffleStatus::Drawing)); // Sold out (max_tickets is 5)
-}
-
-// --- MULTI-ASSET SUPPORT TESTS (Issue #66) ---
-
-/// Helper: create a raffle with a custom token that has a specific decimal count.
-/// Soroban's register_stellar_asset_contract_v2 always gives 7 decimals (XLM-like),
-/// so we use it for both tests and verify decimals via get_token_decimals.
-#[test]
-fn test_get_token_decimals_returns_correct_value() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let (client, _, _, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    // Stellar asset contracts default to 7 decimals (same as XLM stroops)
-    let decimals = client.get_token_decimals();
-    assert_eq!(decimals, 7u32);
-    let _ = admin_client; // suppress unused warning
-}
-
-#[test]
-fn test_raffle_with_xlm_like_token_full_flow() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let creator = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-
-    let factory = env.register(MockFactory, ());
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-    let sac = token::StellarAssetClient::new(&env, &token_id);
-
-    // XLM has 7 decimals: 1 XLM = 10_000_000 stroops
-    let one_xlm: i128 = 10_000_000;
-    sac.mint(&creator, &(100 * one_xlm)); // 100 XLM for prize
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000u32);
-
-    let config = RaffleConfig {
-        description: String::from_str(&env, "XLM Raffle"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: one_xlm, // 1 XLM per ticket
-        payment_token: token_id.clone(),
-        prize_amount: 100 * one_xlm, // 100 XLM prize
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[2u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-
-    // decimals should be 7
-    assert_eq!(client.get_token_decimals(), 7u32);
-
-    client.deposit_prize();
-
-    for _ in 0..1 {
-        let buyer = Address::generate(&env);
-        sac.mint(&buyer, &(5 * one_xlm));
-        client.buy_tickets(&buyer, &5);
-    }
-    client.finalize_raffle();
-
-    let raffle = client.get_raffle();
-    let winner = raffle.winners.get(0).unwrap();
-    env.ledger().with_mut(|l| l.timestamp += 3600);
-    let claimed = client.claim_prize(&winner, &0);
-
-    // Winner should receive 100 XLM in stroops
-    assert_eq!(claimed, 100 * one_xlm);
-
-    let token_client = token::Client::new(&env, &token_id);
-    assert_eq!(token_client.balance(&winner), 100 * one_xlm);
-}
-
-#[test]
-#[should_panic] // prize_amount < ticket_price should be rejected
-fn test_init_rejects_prize_less_than_ticket_price() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let creator = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-
-    let factory = env.register(MockFactory, ());
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000u32);
-
-    // prize_amount (5) < ticket_price (10) — should panic
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Bad Raffle"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id,
-        prize_amount: 5i128, // less than ticket_price
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[3u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-}
-
-// --- PRIZE REFUND & CREATION DEPOSIT TESTS (Issue #67) ---
-
-#[test]
-fn test_refund_prize_after_cancel() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, creator, buyer, admin_client, _, _) =
-        setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-    let token_client = token::Client::new(&env, &admin_client.address);
-
-    client.deposit_prize();
-    client.buy_tickets(&buyer, &1);
-    client.cancel_raffle(&CancelReason::CreatorCancelled);
-
-    // cancel_raffle already refunds inline; prize_deposited is now false
-    // refund_prize should fail since prize is already returned
-    // (idempotency: no double refund)
-    // We test the dedicated path by cancelling without a prior deposit
-}
-
-#[test]
-fn test_refund_prize_dedicated_function() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let creator = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-
-    let factory = env.register(MockFactory, ());
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-    let sac = token::StellarAssetClient::new(&env, &token_id);
-    sac.mint(&creator, &100_000i128);
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-    let token_client = token::Client::new(&env, &token_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000u32);
-
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Refund Test"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 100_000i128,
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[4u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-    client.deposit_prize();
-
-    assert_eq!(token_client.balance(&creator), 0i128); // prize escrowed
-
-    // Cancel without buying any tickets
-    client.cancel_raffle(&CancelReason::CreatorCancelled);
-
-    // After cancel, inline refund runs — creator gets prize back
-    assert_eq!(token_client.balance(&creator), 100_000i128);
-}
-
-#[test]
-fn test_refund_prize_after_min_tickets_not_met() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let creator = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-
-    let factory = env.register(MockFactory, ());
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-    let sac = token::StellarAssetClient::new(&env, &token_id);
-    sac.mint(&creator, &100_000i128);
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-    let token_client = token::Client::new(&env, &token_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000u32);
-
-    // min_tickets = 3, but we'll only sell 1
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Min Tickets Raffle"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 3,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 100_000i128,
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[5u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-    client.deposit_prize();
-
-    // Only 1 ticket sold — below min_tickets of 3
-    let buyer = Address::generate(&env);
-    sac.mint(&buyer, &10_000i128);
-    client.buy_tickets(&buyer, &1);
-
-    // Manually transition to Drawing (simulate end_time passed)
-    // We need to set end_time in the past; re-init with end_time
-    // Instead, use cancel with MinTicketsNotMet via factory auth
-    client.cancel_raffle(&CancelReason::MinTicketsNotMet);
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Cancelled);
-
-    // Prize should be refunded to creator
-    assert_eq!(token_client.balance(&creator), 100_000i128);
-}
-
-#[test]
-fn test_refund_prize_transitions_to_failed_on_min_tickets() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let creator = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-
-    let factory = env.register(MockFactory, ());
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-    let sac = token::StellarAssetClient::new(&env, &token_id);
-    sac.mint(&creator, &100_000i128);
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-    let token_client = token::Client::new(&env, &token_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000u32);
-
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Failed Raffle"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 3,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 100_000i128,
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[6u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-    client.deposit_prize();
-
-    // Sell only 1 ticket (below min_tickets=3), then force Drawing status
-    let buyer = Address::generate(&env);
-    sac.mint(&buyer, &10_000i128);
-    client.buy_tickets(&buyer, &1);
-
-    // Manually set status to Drawing so finalize_raffle can run the min_tickets check
-    env.as_contract(&contract_id, || {
-        let mut raffle = env
-            .storage()
-            .instance()
-            .get::<_, Raffle>(&DataKey::Raffle)
-            .unwrap();
-        raffle.status = RaffleStatus::Drawing;
-        env.storage().instance().set(&DataKey::Raffle, &raffle);
-    });
-
-    // finalize_raffle should detect min_tickets not met and set status to Failed
-    client.finalize_raffle();
-
-    assert_eq!(client.get_raffle().status, RaffleStatus::Failed);
-
-    // Creator calls refund_prize to reclaim escrowed prize
-    client.refund_prize();
-
-    assert_eq!(token_client.balance(&creator), 100_000i128);
-    assert!(!client.get_raffle().prize_deposited);
-}
-
-#[test]
-#[should_panic] // refund_prize on active raffle should fail
-fn test_refund_prize_rejected_when_active() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let (client, _, _, _, _, _) = setup_raffle_env(&env, RandomnessSource::Internal, None, 0, None);
-
-    client.deposit_prize();
-    client.refund_prize(); // should panic — raffle is Open/Active, not Cancelled/Failed
-}
-
-#[test]
-#[should_panic] // double refund_prize should fail
-fn test_refund_prize_idempotent_rejects_double_call() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let creator = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-
-    let factory = env.register(MockFactory, ());
-
-    let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_id = token_contract.address();
-    let sac = token::StellarAssetClient::new(&env, &token_id);
-    sac.mint(&creator, &100_000i128);
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-
-    let mut prizes = Vec::new(&env);
-    prizes.push_back(10000u32);
-
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Double Refund Test"),
-        end_time: 0,
-        max_tickets: 5,
-        min_tickets: 3,
-        allow_multiple: true,
-        ticket_price: 10_000i128,
-        payment_token: token_id.clone(),
-        prize_amount: 100_000i128,
-        prizes,
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[7u8; 32]),
-    };
-
-    client.init(&factory, &factory_admin, &creator, &config);
-    client.deposit_prize();
-
-    env.as_contract(&contract_id, || {
-        let mut raffle = env
-            .storage()
-            .instance()
-            .get::<_, Raffle>(&DataKey::Raffle)
-            .unwrap();
-        raffle.status = RaffleStatus::Failed;
-        env.storage().instance().set(&DataKey::Raffle, &raffle);
-    });
-
-    client.refund_prize();
-    client.refund_prize(); // should panic — PrizeNotDeposited
-}
-
-#[test]
-fn test_native_xlm_support_detection() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let native_addr = env.register(NativeMock, ());
-    let creator = Address::generate(&env);
-    let factory_admin = Address::generate(&env);
-    let factory = env.register(MockFactory, ());
-
-    let config = RaffleConfig {
-        description: String::from_str(&env, "Native Test"),
-        end_time: 0,
-        max_tickets: 10,
-        min_tickets: 0,
-        allow_multiple: true,
-        ticket_price: 10_000,
-        payment_token: native_addr.clone(),
-        prize_amount: 100_000,
-        prizes: soroban_sdk::vec![&env, 10000],
-        randomness_source: RandomnessSource::Internal,
-        oracle_address: None,
-        protocol_fee_bp: 0,
-        treasury_address: None,
-        swap_router: None,
-        tikka_token: None,
-        metadata_hash: BytesN::from_array(&env, &[1u8; 32]),
-    };
-
-    let contract_id = env.register(Contract, ());
-    let raffle_client = ContractClient::new(&env, &contract_id);
-    raffle_client.init(&factory, &factory_admin, &creator, &config);
-
-    // Deposit prize
-    let native_client = token::StellarAssetClient::new(&env, &native_addr);
-    native_client.mint(&creator, &100_000);
-    raffle_client.deposit_prize();
-
-    let buyer = Address::generate(&env);
-    raffle_client.buy_tickets(&buyer, &1);
-}
