--- contracts/raffle/src/oracle.rs
@@ -1,42 +1,0 @@
-use soroban_sdk::{contractclient, contracttype, Address, Env};
-
-/// All context needed for a round-trip randomness request/response.
-/// Passed by value to any `RandomnessOracleTrait` implementor so the oracle
-/// knows where to deliver the result and which request it is fulfilling.
-#[derive(Clone)]
-#[contracttype]
-pub struct RandomnessRequest {
-    /// The raffle contract address that is requesting randomness.
-    pub raffle_id: Address,
-    /// Unique identifier for this request — set to `ledger.sequence()` at
-    /// dispatch time so it is deterministic and non-repeating per ledger.
-    pub request_id: u64,
-    /// The contract address the oracle must call back with the result.
-    pub callback_address: Address,
-}
-
-/// Interface every oracle provider must implement.
-///
-/// Any contract deployed as an oracle (Acurast, future Stellar VRF, etc.)
-/// must expose this function. The raffle calls it via `RandomnessOracleClient`
-/// generated by the `contractclient` macro.
-#[contractclient(name = "RandomnessOracleClient")]
-pub trait RandomnessOracleTrait {
-    /// Request a random value. The oracle is expected to call back
-    /// `receive_randomness` on `request.callback_address` asynchronously.
-    fn request_randomness(env: Env, request: RandomnessRequest);
-}
-
-/// Callback interface the raffle contract exposes to oracle providers.
-///
-/// An oracle that has fulfilled a `RandomnessRequest` calls this on the
-/// raffle contract via `RandomnessReceiverClient`. The raffle validates
-/// the `request_id` and finalises the draw.
-#[contractclient(name = "RandomnessReceiverClient")]
-pub trait RandomnessReceiverTrait {
-    /// Deliver a random seed back to the raffle.
-    ///
-    /// - `request_id`: must match the pending request stored during dispatch
-    /// - `random_seed`: the verifiable random value produced by the oracle
-    fn receive_randomness(env: Env, request_id: u64, random_seed: u64);
-}
